requirements.txt
aiogoogle==1.1.14
pillow==8.3.1